🍟

+++ initWizard.sh
@@ -8,24 +8,15 @@ Yellow='\e[0;33m'       # Yellow
 Blue='\e[0;34m'         # Blue
 Purple='\e[0;35m'       # Purple
 Cyan='\e[0;36m'         # Cyan
-White='\e[0;37m'        # White
+White='\e[0m'           # White
 Grey='\e[0;90m'         # Grey
 
 # Bold
-BBlack='\e[1;30m'       # Black
-BRed='\e[1;31m'         # Red
-BGreen='\e[1;32m'       # Green
-BYellow='\e[1;33m'      # Yellow
 BBlue='\e[1;34m'        # Blue
-BPurple='\e[1;35m'      # Purple
-BCyan='\e[1;36m'        # Cyan
-BWhite='\e[1;37m'       # White
-BGrey='\e[1;90m'        # Grey
 
 choose_from_menu() {
     local prompt="$1" outvar="$2"
-    shift
-    shift
+    shift 2
     local options=("$@") cur=0 count=${#options[@]} index=0
     local esc=$(echo -en "\e") # cache ESC as test doesn't allow esc codes
     local max_length=0
@@ -37,11 +28,11 @@ choose_from_menu() {
         fi
     done
     
-      # Hide cursor
+    # Hide cursor
     echo -en "${esc}[?25l"
 
     # Trap SIGINT (Ctrl+C) and show cursor before exiting
-    trap 'echo -en "${esc}[?25h"; exit 1' INT
+    trap 'echo -e "${Red}✖${White} Operation cancelled";echo -en "${esc}[?25h"; exit 1' INT
 
     while true; do
         # Display the menu prompt
@@ -84,7 +75,7 @@ choose_from_menu() {
     trap - INT
 
     # Display the selected option with a ✔
-    echo -e "${Green}✔ ${White}${outvar} ${Grey}› ${Purple}${options[$cur]}"
+    echo -e "${Green}✔ ${White}${prompt} ${Grey}› ${Purple}${options[$cur]}${White}"
     # Clear the rest of the screen
     tput ed
 }
@@ -114,31 +105,32 @@ directory_exists() {
 log_message() {
     local type="$1"
     local message="$2"
-    local timestamp=$(date +"%Y-%m-%d  %H:%M:%S")
+    local timestamp=$(date +"%Y-%m-%d %H:%M:%S")
     local username=$(whoami)
-   sudo echo "$timestamp : $username : $type : $message" | tee -a /var/log/initWizard/history.log
+    echo "$timestamp : $username : $type : $message" | sudo tee -a /var/log/initWizard/history.log
 }
 
 # Function to execute the program in a subshell
 execute_in_subshell() {
     local script="$1"
+    local args=("$@")  # Store all arguments in an array
     # Execute the script in a subshell
     (
-        source "$script"
+        source "$script" "${args[@]}"
     )
 }
 
 # Function to display help message
 display_help() {
-    echo -e "${BWhite}Usage:${White} $0 [options] [paramètre]\n"
-    echo -e "${BWhite}Options:${White}"
+    echo -e "${White}Usage:${White} $0 [options]\n"
+    echo -e "${White}Options:"
     echo -e "  ${BBlue}-h${White}, ${BBlue}--help${White}          Display this help message"
     echo -e "  ${BBlue}-f${White}, ${BBlue}--fork${White}          Enable fork execution"
     echo -e "  ${BBlue}-t${White}, ${BBlue}--thread${White}        Enable thread execution"
-    echo -e "  ${BBlue}-s${White}, ${BBlue}--subshell${White}     Execute program in a subshell"
+    echo -e "  ${BBlue}-s${White}, ${BBlue}--subshell${White}      Execute program in a subshell"
     echo -e "  ${BBlue}-l${White}, ${BBlue}--log${White}           Specify a directory for log file storage"
     echo -e "  ${BBlue}-r${White}, ${BBlue}--restore${White}       Reset to default settings (admin only)"
-    echo -e "\n${BWhite}Example:${White} $0 --log"
+    echo -e "\n${White}Example:${White} $0 --log"
 }
 
 # Function to initialize a git repository
@@ -156,6 +148,24 @@ create_readme() {
     echo "# $project_name" > "$project_name/README.md"
 }
 
+# Function to display the banner
+display_banner() {
+    echo ""
+    sleep 0.1
+    echo -e "   ${Green}██╗███╗   ██╗██╗████████╗    ${Red}██${Grey}╗    ${Red}██${Grey}╗${Red}██${Grey}╗${Red}███████${Grey}╗ ${Red}█████${Grey}╗ ${Red}██████${Grey}╗ ${Red}██████${Grey}╗"
+    sleep 0.1
+    echo -e "   ${Green}██║████╗  ██║██║╚══██╔══╝    ${Red}██${Grey}║    ${Red}██${Grey}║${Red}██${Grey}║╚══${Red}███${Grey}╔╝${Red}██${Grey}╔══${Red}██${Grey}╗${Red}██${Grey}╔══${Red}██${Grey}╗${Red}██${Grey}╔══${Red}██${Grey}╗"
+    sleep 0.1
+    echo -e "   ${Green}██║██╔██╗ ██║██║   ██║       ${Red}██${Grey}║ ${Red}█${Grey}╗ ${Red}██${Grey}║${Red}██${Grey}║  ${Red}███${Grey}╔╝ ${Red}███████${Grey}║${Red}██████${Grey}╔╝${Red}██${Grey}║  ${Red}██${Grey}║"
+    sleep 0.1
+    echo -e "   ${Green}██║██║╚██╗██║██║   ██║       ${Red}██${Grey}║${Red}███${Grey}╗${Red}██${Grey}║${Red}██${Grey}║ ${Red}███${Grey}╔╝  ${Red}██${Grey}╔══${Red}██${Grey}║${Red}██${Grey}╔══${Red}██${Grey}╗${Red}██${Grey}║  ${Red}██${Grey}║"
+    sleep 0.1
+    echo -e "   ${Green}██║██║ ╚████║██║   ██║       ${Grey}╚${Red}███${Grey}╔${Red}███${Grey}╔╝${Red}██${Grey}║${Red}███████${Grey}╗${Red}██${Grey}║  ${Red}██${Grey}║${Red}██${Grey}║  ${Red}██${Grey}║${Red}██████${Grey}╔╝"
+    sleep 0.1
+    echo -e "   ${Green}╚═╝╚═╝  ╚═══╝╚═╝   ╚═╝        ${Grey}╚══╝╚══╝ ╚═╝╚══════╝╚═╝  ╚═╝╚═╝  ╚═╝╚═════╝"
+    echo -e "${White}"
+}
+
 if [ "$(id -u)" -ne 0 ]; then
     echo "This script must be run as root" >&2
     exit 1
@@ -164,7 +174,7 @@ fi
 # Check if log directory exists, if not, create it
 log_dir="/var/log/InitWizard"
 if ! directory_exists "$log_dir"; then
-   sudo mkdir -p "$log_dir"
+    sudo mkdir -p "$log_dir"
 fi
 
 # Check if log file exists, if not, create it
@@ -188,7 +198,7 @@ while [[ $# -gt 0 ]]; do
             ;;
         -s|--subshell)
             # Execute the program in a subshell
-            execute_in_subshell "$0"
+            execute_in_subshell "$0" "$@"
             exit 0
             ;;
         *)
@@ -199,48 +209,73 @@ while [[ $# -gt 0 ]]; do
     esac
 done
 
-# Function to display the banner
-display_banner() {
-    echo ""
-    sleep 0.1
-    echo -e "   ${Green}██╗███╗   ██╗██╗████████╗    ${Red}██${Grey}╗    ${Red}██${Grey}╗${Red}██${Grey}╗${Red}███████${Grey}╗ ${Red}█████${Grey}╗ ${Red}██████${Grey}╗ ${Red}██████${Grey}╗"
-    sleep 0.1
-    echo -e "   ${Green}██║████╗  ██║██║╚══██╔══╝    ${Red}██${Grey}║    ${Red}██${Grey}║${Red}██${Grey}║╚══${Red}███${Grey}╔╝${Red}██${Grey}╔══${Red}██${Grey}╗${Red}██${Grey}╔══${Red}██${Grey}╗${Red}██${Grey}╔══${Red}██${Grey}╗"
-    sleep 0.1
-    echo -e "   ${Green}██║██╔██╗ ██║██║   ██║       ${Red}██${Grey}║ ${Red}█${Grey}╗ ${Red}██${Grey}║${Red}██${Grey}║  ${Red}███${Grey}╔╝ ${Red}███████${Grey}║${Red}██████${Grey}╔╝${Red}██${Grey}║  ${Red}██${Grey}║"
-    sleep 0.1
-    echo -e "   ${Green}██║██║╚██╗██║██║   ██║       ${Red}██${Grey}║${Red}███${Grey}╗${Red}██${Grey}║${Red}██${Grey}║ ${Red}███${Grey}╔╝  ${Red}██${Grey}╔══${Red}██${Grey}║${Red}██${Grey}╔══${Red}██${Grey}╗${Red}██${Grey}║  ${Red}██${Grey}║"
-    sleep 0.1
-    echo -e "   ${Green}██║██║ ╚████║██║   ██║       ${Grey}╚${Red}███${Grey}╔${Red}███${Grey}╔╝${Red}██${Grey}║${Red}███████${Grey}╗${Red}██${Grey}║  ${Red}██${Grey}║${Red}██${Grey}║  ${Red}██${Grey}║${Red}██████${Grey}╔╝"
-    sleep 0.1
-    echo -e "   ${Green}╚═╝╚═╝  ╚═══╝╚═╝   ╚═╝        ${Grey}╚══╝╚══╝ ╚═╝╚══════╝╚═╝  ╚═╝╚═╝  ╚═╝╚═════╝"
-    echo -e "${White}"
-}
-
 display_banner
 
-echo -en "${Blue}? ${White}Project name: "
+# Trap SIGINT (Ctrl+C) and show a message before exiting
+trap 'echo -e "\n${Red}✖${White} Operation cancelled"; exit 1' INT
+
+echo -en "${Blue}? ${White}Project name: ${Grey}›${White} "
 read project_name
+
+# if user presses enter without entering a project name
+if [ -z "$project_name" ]; then
+    # Clear the line where the project name prompt was displayed
+    echo -en "\033[1A"  # Move cursor up one line
+    echo -en "\r\033[K"  # Clear the line
+
+    echo -e "${Red}✖${White} Project name cannot be empty"
+    echo -e "${Red}✖${White} Operation cancelled"
+    exit 1
+fi
 
 # Check if the project folder already exists
 m=true
 while $m ; do
     if [ -d "$project_name" ]; then
-        echo -e "${Red}Error:${White} Project folder '$project_name' already exists."
-        read -p "Do you want to replace it (r) or make a new one with a different name (n)? " choice
+        # Clear the line where the project name prompt was displayed
+        echo -en "\033[1A"  # Move cursor up one line
+        echo -en "\r\033[K"  # Clear the line
+
+        echo -e "${Red}✖${White} Project folder ${Yellow}'$project_name'${White} already exists."
+        echo -en "Do you want to replace it (${Cyan}r${White}) or make a new one with a different name (${Green}n${White})? "
+        read choice
+
+        # Clear the two lines where the options were displayed
+        echo -en "\033[1A"   # Move cursor up one line
+        echo -en "\r\033[K"  # Clear the line
+        echo -en "\033[1A"   # Move cursor up one line
+        echo -en "\r\033[K"  # Clear the line
+
         case "$choice" in
-            [Rr]* ) rm -rf "$project_name";;
-            [Nn]* ) read -p "Enter a new project name: " project_name;;
-            * ) echo "Invalid option. Exiting."; exit 1;;
+            [Rr]* ) rm -rf "$project_name"
+                    echo ""
+                    ;;
+            [Nn]* ) echo -en "${Blue}? ${White}Project name: ${Grey}›${White} "
+                    read project_name
+                    ;;
+            * ) echo -e "${Red}✖${White} Invalid option"
+                ;;
         esac
     else
         m=false
     fi
 done
 
+# Remove the SIGINT trap
+trap - INT
+
+# Clear the line where the project name prompt was displayed
+echo -en "\033[1A"   # Move cursor up one line
+echo -en "\r\033[K"  # Clear the line
+
+# Display the selected project name with a ✔
+echo -e "${Green}✔ ${White}Project name: ${Grey}› ${Purple}${project_name}"
+
+# Display the language selection menu
 languages=("C" "Cpp" "Node" "Vite")
 choose_from_menu "Select a language:" language "${languages[@]}"
 
+# Display the git repository creation menu
 selections=("Yes" "No")
 choose_from_menu "Do you want to create a git repository?" git "${selections[@]}"
